dev-requirements.txt
pylint==3.2.1
httpretty==1.1.4
mypy==1.9.0
sphinx==7.1.2
sphinx-rtd-theme==2.0.0rc4
sphinx-autodoc-typehints==1.25.2
pytest==7.4.4
pytest-cov==4.1.0
readme-renderer==42.0
markupsafe==2.1.3
bleach==4.1.0 # This dependency was updated to a breaking version.
codespell==2.1.0
requests==2.32.3
ruamel.yaml==0.17.21
asgiref==3.7.2
psutil==5.9.6
GitPython==3.1.41
pre-commit==3.7.0; python_version >= '3.9'
pre-commit==3.5.0; python_version < '3.9'
ruff==0.6.9
